refactor: axios client
