+++ src/modules/core/errors/trackers.ts
@@ -2,6 +2,7 @@ import { HttpAdapterHost, BaseExceptionFilter } from '@nestjs/core';
 import { INestApplication } from '@nestjs/common';
 import { init as initSentry, captureException as captureOnSentry, captureConsoleIntegration, setupNestErrorHandler as setupSentryNestErrorHandler } from '@sentry/nestjs';
 import { nodeProfilingIntegration } from '@sentry/profiling-node';
+import { GraphQLFormattedError } from 'graphql';
 import readPackageInfo from '@common/utils/packageInfoReader.util';
 import { EnvironmentsEnum } from '@common/enums/environments.enum';
 
@@ -34,4 +35,17 @@ export const configureTrackers = (
 		release: `${packageName}@${packageVersion}`,
 	});
 	setupSentryNestErrorHandler(nestApp, new BaseExceptionFilter(httpAdapterHost.httpAdapter));
+};
+
+export const formatGraphQlError = ({ message, extensions, path }: GraphQLFormattedError, error: any): GraphQLFormattedError => {
+	const graphQLFormattedError: GraphQLFormattedError = {
+		message: message ?? error?.message,
+		path: path ?? error?.path,
+		extensions: {
+			code: extensions?.code,
+			originalError: extensions?.originalError,
+		},
+	};
+
+	return graphQLFormattedError;
 };
