+++ src/modules/core/infra/providers/RestMockedService.provider.ts
@@ -13,12 +13,11 @@ export default class RestMockedServiceProvider extends AbstractRestClient {
 		configService: ConfigService,
 		logger: LoggerService,
 	) {
-		const { baseUrl, serviceName, timeout } = configService.get<ConfigsInterface['integration']['rest']['mockedService']>('integration.rest.mockedService')!;
+		const { baseUrl, serviceName, timeout, maxRetries, maxRedirects } = configService.get<ConfigsInterface['integration']['rest']['mockedService']>('integration.rest.mockedService')!;
 
 		super({
-			serviceName,
-			baseUrl,
-			timeout,
+			serviceName, baseUrl,
+			timeout, maxRedirects, maxRetries,
 			logger,
 		});
 	}
